requirements.txt
fastapi
uvicorn
python-dotenv
docx2txt
PyPDF2
pinecone-client
langchain-community
langchain-pinecone
langchain-groq
langchain
pydantic
python-multipart